langgraph_run/requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0

# Google Vertex AI
google-cloud-aiplatform>=1.60.0
langchain-google-vertexai>=2.0.0

# Audio processing
openai-whisper>=20231117
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Utility packages
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0